% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodymapr_df.R
\name{bodymapr_df}
\alias{bodymapr_df}
\title{Create a data frame of clinical tumor characteristics from a structured electronic data collection instrument (e.g. the Lesion Information instrument) and maps them to a look up table that contains a coordinate system for the BodyMapR}
\usage{
bodymapr_df(data)
}
\arguments{
\item{data}{is a data frame which contains the data for which you want to create a body map. This data is typically an exported csv from a REDCap project that has incorporated the Lesion Information Instrument.}
}
\value{
Data Frame
}
\description{
`bodymapr_df()` returns a data frame of lesions from a electronic data capture project that uses the Lesion Information Instrument. This is a intermediary step for the body_map.plot() function.
}
\examples{
# Test with embedded data set "mock_mcc_dataset"
BodyMapR::mock_mcc_dataset \%>\%
  bodymapr.df()

}
