% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodymapr_plot.R
\name{bodymapr_plot}
\alias{bodymapr_plot}
\title{Create an interactive data visualization of clinical tumor characteristics from a structured EDC lesion instrument for a Shiny App}
\usage{
bodymapr_plot(data)
}
\arguments{
\item{data}{is a data frame which contains the data for which you want to create a body map after it has been processed for this application}
}
\value{
Plotly Visualization
}
\description{
`bodymapr_plot()`returns a plotly visualization of cutaneous tumor characteristics plotted on a body map
}
\examples{
# Test with embedded data set "mock_dataset"
BodyMap::mock_mcc_dataset \%>\%
  body_map_plot_shiny()

}
