% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics.df.unite.R
\name{genomics.df.unite}
\alias{genomics.df.unite}
\title{Create a data frame with genomic alterations listed in wide format with all genomic alterations in one cell}
\usage{
genomics.df.unite(data)
}
\arguments{
\item{data}{data frame/tibble}
}
\value{
A data frame with the list of genes altered in a single cell
}
\description{
`genomics.df.unite()` wrangles and processes genomics data from a REDCap project that has incorporated the Genomics Instrument. This allows for expedited analysis of patient-level data from REDCap. Genetic alterations are listed in wide format with a concatenation of genomic alterations in one cell.
}
\examples{
# Create a new DF with sample data set
BodyMap::mock_mcc_dataset \%>\%
  genomics.df.unite()

}
